backend/src/requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.44.0
django-appconf==1.0.5
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-environ==0.7.0
django-extensions==3.1.3
django-image-cropping==1.6.1
django-resized==0.3.11
django-storages==1.11.1
django-taggit==1.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.19.0
drf-yasg==1.20.0
easy-thumbnails==2.7.1
factory-boy==3.2.0
Faker==8.14.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
Werkzeug==2.0.1
